refactor: index.js를 통해 컴포넌트에 접근하도록 변경
