+++ src/pages/Baskets/index.js
@@ -1,0 +1,1 @@
+export { default } from './BasketsPage';
