+++ src/pages/Product/index.js
@@ -1,0 +1,1 @@
+export { default } from './ProductPage';
